Billboard click detection fix

+++ all/native/renderers/drawdatas/BillboardDrawData.h
@@ -74,8 +74,8 @@ namespace carto {
         double getCameraPlaneZoomDistance() const;
         void setCameraPlaneZoomDistance(double cameraPlaneDistance);
     
-        double getScreenBottomDistance() const;
-        void setScreenBottomDistance(double screenBottomDistance);
+        int getScreenBottomDistance() const;
+        void setScreenBottomDistance(int screenBottomDistance);
     
         BillboardRenderer* getRenderer() const;
         void setRenderer(BillboardRenderer& renderer);
@@ -135,7 +135,7 @@ namespace carto {
         float _size;
     
         double _cameraPlaneZoomDistance;
-        double _screenBottomDistance;
+        int _screenBottomDistance;
     
         BillboardRenderer* _renderer;
     };
